Add files via upload
introduced pellets and power pellets to the board. Pellets are the fruits that Pacman has to eat to win each round. Also added the power feature that allows the packman to eat the ghost. Will look to find a way to move the movement options to the movement.py file